Added Helper Methods
StructureGenHelper is now fulfilling its main purpose. Lots of new
helper methods now exist which allow a player to create a structure with
the new fast set block method. The input is the world object, a 3D array
of BlockData (new helper class), and the starting coordinates for the
structure.