Step 1 of Safe Area Block
The rendering semi-works. I just have to figure out how to tie in the
block's "should side be rendered" with the renderer's "draw the block's
face." It seems like I need data from both. This is *not* finished yet,
I am going to be modifying the safe area block and its renderer more.