Added Slabs/Stairs and Finished Safe Area Block!
The renderer for the safe area block was a bit of difficult to figure
out, but I finally found some vanilla code that was similar in function
to what I wanted to do with the block. I used that and modified it. For
the stairs, it was easy. Just set the icon correctly. For the slabs, I
had to do a bit of work and add extra classes but it functions and
renders like a normal slab now.