Unsheather class and KagemitsuG4
This commit adds the KagemitsuG4 to compliment the Object Eraser, and
adds an abstract class for swords that like to fwoosh out at various
speeds. All of it is adjustable to a degree. It would need to be edited
a bit and might become slow if it got more precise speed tweaks.