Added in 100 Floors as Templates
Floors are now represented in data as Floor Points, which store lots of
data about how the floor gets generated in world gen. There are a few
helper methods to make sure that changing one of them excessively does
not mess anything up. These starter points and their data are then used
to generate the floors in world gen. We can also use them to add the
towers and dungeons in, based on the center point and the size of the
floors.